Operating Systems (CS446)/Assignment4/jeffreycarson_CS446_PA4/requirements.txt
numpy == 1.21.2
matplotlib == 3.5.1
opencv-python == 4.5.5.64
pyOpenSSL == 22.0.0
